feat(web): use plan id instead of order id for status page
The reason behind this change is that we do not have access to the
`orderId` after the purchase action. Thus, it would require further
development to retrieve that value and use it on the status page.

The solution is to use the `planId` instead and check if the user has an
active order with that `planId`. The user would of course need to log in
to assert that they are owners of the subscription for that plan.

Support for this new approach will be developed in follow-up commits